Categorization Test Dataframe
Note: This dataframe is not suitable for actual algorithm training. The non-ira tweets seem vastly less political than the ira tweets so that's what it's more likely to sort by. This is just to check if a classifier can be made that sorts tweets according to text into ira or non-ira categories and if that proves successful, we will get a better dataframe for training data.